% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chgg4.R
\name{chgg4}
\alias{chgg4}
\title{GG4 Cumulative Hazard Function}
\usage{
chgg4(x, a, l, c, mu = 0)
}
\arguments{
\item{x}{data}

\item{a}{scale parameter}

\item{l}{shape parameter}

\item{c}{power shape parameter}

\item{mu}{location parameter}
}
\value{
Cumulative hazard ratio
}
\description{
GG4 CHF: Cumulative hazard function of the four-parameter Generalized gamma distribution
}
\examples{
## a = 2, l = 3, c = 4 and mu = 5
rr = rgg4(100,2,3,4,5,sequence = TRUE)
c(mean(rr),sd(rr))
## [1] 7.5497902 0.3718677
q = sort(rr)
chh = chgg4(q,2,3,4,5)
summary(chh)
## Min.    1st Qu.  Median    Mean    3rd Qu.  Max.
## 0.00995 0.29432  0.69320   0.97773 1.36683  4.61512

}
\seealso{
\code{\link{dgg4}},  \code{\link{pgg4}}, \code{\link{qgg4}}, \code{\link{rgg4}},\code{\link{hgg4}}
}
