% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddhplot.R
\name{ddhplot}
\alias{ddhplot}
\title{Densities and histogramm observations versus model GG4 estimated}
\usage{
ddhplot(y, theta, seed = 986)
}
\arguments{
\item{y}{the observed sample}

\item{theta}{alues of parameters.}

\item{seed}{seed initialization}
}
\value{
a QQ-plot, densities and hystogramm of observations vs estimated model of GG4
}
\description{
Densities and histogramm observations versus model GG4 estimated
}
