% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityplot.R
\name{densityplot}
\alias{densityplot}
\title{Density plots observations versus model GG4 estimated}
\usage{
densityplot(y, theta)
}
\arguments{
\item{y}{the observed sample}

\item{theta}{Vector of salues of the four-parameters.}
}
\value{
a QQ-plot, densities and hystogramm of observations vs estimated model of GG4
}
\description{
Density plots observations versus model GG4 estimated
}
