% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgg4.R
\name{dgg4}
\alias{dgg4}
\title{GG4 Probability Density Function}
\usage{
dgg4(x, a, l, c, mu = 0, log.p = F)
}
\arguments{
\item{x}{vector of quantile}

\item{a}{scale parameter}

\item{l}{shape parameter}

\item{c}{power shape parameter}

\item{mu}{location parameter}

\item{log.p}{Logical; if TRUE, probabilities p are given as log(p)}
}
\value{
density
}
\description{
GG4 PDF: Probability Density Function of the four-parameter Generalized Gamma Distribution
}
\examples{
set.seed(963)
rr = rgg4(100,2,3,4,5,sequence = FALSE)
c(mean(rr),sd(rr))
## [1] 7.505230 0.368037
q = sort(rr)
dd = dgg4(q,2,3,4,5)
plot(dd)

}
\seealso{
\code{\link{pgg4}}, \code{\link{qgg4}}, \code{\link{rgg4}},\code{\link{hgg4}}, \code{\link{chgg4}}
}
