% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.mjse.R
\name{fit.mjse}
\alias{fit.mjse}
\title{MJSE: Minimum Jensen-Shanon Bregman Divergence Estimate based on pdf for the 4-parameter GG4 distribution}
\usage{
fit.mjse(y, theta)
}
\arguments{
\item{y}{the observed sample}

\item{theta}{initial values of parameters.}
}
\value{
a vector of four-parameter estimated according observation sample + info on convergence
}
\description{
Fit of univariate distributions to non-censored data by minimum Jensen-Shanon Bregman divergence based on pdf.
}
\details{
See R code example in vignettes directory
}
\seealso{
\code{\link{fit.mle}},
\code{\link{fit.mkle}},
\code{\link{fit.mhe}},
\code{\link{fit.msqe}},
\code{\link{fit.mwe}},
\code{\link{fit.mhdfe}},
\code{\link{fit.mwdfe}},
\code{\link{fit.msqdfe}}.
}
