% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.mle.R
\name{fit.mle}
\alias{fit.mle}
\title{MLE: Maximum Likelihood Estimate of the 4-parameter GG4 disrtibution}
\usage{
fit.mle(y, theta)
}
\arguments{
\item{y}{the observed sample}

\item{theta}{initial values of parameters.}
}
\value{
a vector of four-parameter estimated according observation sample + info on convergence
}
\description{
Fit of univariate distributions to non-censored data by maximum likelihood (mle)
}
\details{
See R code example in vignettes directory
}
\seealso{
\code{\link{fit.mkle}},
\code{\link{fit.mjse}},
\code{\link{fit.mhe}},
\code{\link{fit.msqe}},
\code{\link{fit.mwe}},
\code{\link{fit.mhdfe}},
\code{\link{fit.mwdfe}},
\code{\link{fit.msqdfe}}.
}
