% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.mwdfe.R
\name{fit.mwdfe}
\alias{fit.mwdfe}
\title{MWDFE: Minimum Wasserstein distance Estimate based on cdf for the 4-parameter GG4 distribution}
\usage{
fit.mwdfe(y, theta, d = 1)
}
\arguments{
\item{y}{the observed sample}

\item{theta}{initial values of parameters.}

\item{d}{exponent of the Wasserstein metric}
}
\value{
a vector of four-parameter estimated according observation sample + info on convergence
}
\description{
Fit of univariate distributions to non-censored data by minimum Wasserstein distance estimate based on cdf.
}
\details{
See R code example in vignettes directory
}
\seealso{
\code{\link{fit.mle}},
\code{\link{fit.mkle}},
\code{\link{fit.mjse}},
\code{\link{fit.mhe}},
\code{\link{fit.msqe}},
\code{\link{fit.mwe}},
\code{\link{fit.mhdfe}},
\code{\link{fit.msqdfe}}.
}
