% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgg4.R
\name{hgg4}
\alias{hgg4}
\title{GG4 Hazard Function}
\usage{
hgg4(x, a, l, c, mu = 0)
}
\arguments{
\item{x}{data}

\item{a}{scale parameter}

\item{l}{shape parameter}

\item{c}{power shape parameter}

\item{mu}{location parameter}
}
\value{
hazard ratio
}
\description{
GG4 HF: Hazard function of the four-parameter Generalized Gamma Distribution
}
\examples{
## a = 2, l = 3, c = 4 and mu = 5
rr = rgg4(100,2,3,4,5,sequence = TRUE)
c(mean(rr),sd(rr))
## [1] 7.5497902 0.3718677
q = sort(rr)
hh = hgg4(q,2,3,4,5)
summary(hh)
## Min.    1st Qu.  Median    Mean    3rd Qu.  Max.
## 0.07809 1.15079  2.06266   2.27508 3.16877  6.60587

}
\seealso{
\code{\link{dgg4}},  \code{\link{pgg4}}, \code{\link{qgg4}}, \code{\link{rgg4}}, \code{\link{chgg4}}
}
