% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.theta.R
\name{init.theta}
\alias{init.theta}
\title{Parameter Initialization}
\usage{
init.theta(
  data,
  lower = -20,
  upper = 20,
  length = 1000,
  a.pos = TRUE,
  p = 0.99,
  q = 1.01
)
}
\arguments{
\item{data}{the observed sample}

\item{lower}{the lower interval value of the discrete values of the power shape parameter $c$ except zero.
Default value: -20}

\item{upper}{the upper interval value of the discrete values of the power shape parameter $c$ except zero.
Default value: +20}

\item{length}{size of the number of discrete values of the power shape parameter $c$ except zero.
Default value: 1,000}

\item{a.pos}{boolean pos = TRUE a>0 else a<0}

\item{p}{coefficient to initialate init.mu for a>0  - Default value: 0.99}

\item{q}{coefficient to initialate init.mu for a<0  - Default value: 1.01}
}
\value{
int.theta a vector of the four-parameter (a,l,c,mu) of GG4 distribution
}
\description{
Iterative method to obtain initial estimates in using the following property: If V follows a two-parameter gamma distribution with scale parameter 1 and shape parameter l, then Y = aV^(1/c)+mu follows a GGD(a,l,c,mu).
}
\details{
In using a sequence of discrete values of the c parameter fixed and  the relation that X^c follows a two-parameter gamma distribution with parameters
a^c and l, the method of moments estimates is used to estimate a^c and l.
Then, for each set of estimates of theta = (a, l, c, mu), we compute the value of the likelihood,
and pick the set that gives the largest likelihood as initial value  four-parameter Generalized Gamma Distribution
So we obtain an initialisation of a vector (a,l,c,mu) as in input for the referenced method (see also paragraph)
See R code examples in vignettes directory.
}
\seealso{
\code{\link{fit.mle}},
\code{\link{fit.mkle}},
\code{\link{fit.mjse}},
\code{\link{fit.mhe}},
\code{\link{fit.msqe}},
\code{\link{fit.mwe}},
\code{\link{fit.mhdfe}},
\code{\link{fit.mwdfe}},
\code{\link{fit.msqdfe}}.
}
