% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtosis.R
\name{kurtosis}
\alias{kurtosis}
\title{Kurtosis of a univariate population}
\usage{
kurtosis(data, type = "sample")
}
\arguments{
\item{data}{a sample (set of observations)}

\item{type}{a character string coding -- sample" is computed from mean and standard deviation
of sample. "fisher" corresponds to fisher'skewness formula.}
}
\value{
K the kurtosis from a sample -- if kurtosis = 0 => Normal distribution
-- kurtosis < 0 => indicates a thin-tailed data distribution (plarkurtic)
-- kurtosis > 0 => indicates a fat-tailed data distribution (lepkurtic)
}
\description{
Kurtosis of a univariate population
}
