% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment_theo.R
\name{moment_theo}
\alias{moment_theo}
\title{Moment estimated from the Generalized Gamma four-parameters. the function
allows us to verify the results regarding parameters estimated with mimic function}
\usage{
moment_theo(theta, k)
}
\arguments{
\item{theta}{vector of 4-parameters.}

\item{k}{order of the moment to be computed from the GGD parameter}
}
\value{
moment from parameters
}
\description{
Moment estimated from the Generalized Gamma four-parameters. the function
allows us to verify the results regarding parameters estimated with mimic function
}
