% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_bias_reduced_estimate.R
\name{mu_bias_reduced_estimate}
\alias{mu_bias_reduced_estimate}
\title{To compute Bias-reduced estimate of mu}
\usage{
mu_bias_reduced_estimate(est_theta, y)
}
\arguments{
\item{est_theta}{list of the 4-parameter GG4 with Est_theta[1]= a, Est_theta[2]=l, Est_theta[3]=C,and Est_theta[4]=mu}

\item{y}{data corresponding to the observed population}
}
\value{
bre_mu  Bias-reduced estimate of mu
}
\description{
To compute Bias-reduced estimate of mu
}
