% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgg4.R
\name{pgg4}
\alias{pgg4}
\title{GG4 Cumulative Distribution Function (cdf)}
\usage{
pgg4(q, a, l, c, mu = 0, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{vector of quantile}

\item{a}{scale parameter}

\item{l}{shape parameter}

\item{c}{power shape parameter}

\item{mu}{location parameter}

\item{lower.tail}{Logical; if TRUE (default), probabilities are P(X≤x), otherwise, P(X>x)}

\item{log.p}{Logical; if TRUE, probabilities p are given as log(p) only if a,l and c are strictly positives}
}
\value{
probability
}
\description{
GG4 CDF: Cumulative distribution function the four-parameter Generalized Gamma Distribution
}
\examples{
# a = 1, l = 2, c = -0.8 and mu = 0 (default value)
p <- (1:9)/10
pgg4(qgg4(p, 1, 2, -0.8), 1, 2, -0.8)
## [1] 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9

}
\seealso{
\code{\link{dgg4}}, \code{\link{qgg4}}, \code{\link{rgg4}},\code{\link{hgg4}}, \code{\link{chgg4}}
}
