% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgg4.R
\name{qgg4}
\alias{qgg4}
\title{GG4 Quantiles/Percentiles generation}
\usage{
qgg4(p, a, l, c, mu = 0, lower.tail = TRUE)
}
\arguments{
\item{p}{vector or quantiles}

\item{a}{scale parameter}

\item{l}{shape parameter}

\item{c}{power shape parameter}

\item{mu}{location parameter}

\item{lower.tail}{indicate lower tail. The default value is set to TRUE}
}
\value{
a vector of quantile generated according to a four parameter GGD
}
\description{
GG4 QPG: Quantiles/Percentiles Generation of the four-parameter Generalized Gamma Distribution
}
\examples{
set.seed(963)
qq = qgg4(runif(100),2,3,4,5)
c(mean(qq),sd(qq))
## [1] 7.505230 0.368037

# ----
# Test
# ----

# a = 1, l = 2, c = -0.8 and mu = 0 (default value)
p <- (1:9)/10
pgg4(qgg4(p, 1, 2, -0.8), 1, 2, -0.8)
## [1] 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9

}
\seealso{
\code{\link{dgg4}}, \code{\link{pgg4}}, \code{\link{rgg4}},\code{\link{hgg4}}, \code{\link{chgg4}}
}
