% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgg4.R
\name{rgg4}
\alias{rgg4}
\title{GG4 Random number generation}
\usage{
rgg4(n, a, l, c, mu = 0, sequence = T)
}
\arguments{
\item{n}{number of observations.}

\item{a}{scale parameter}

\item{l}{shape parameter}

\item{c}{power shape parameter}

\item{mu}{location parameter}

\item{sequence}{if TRUE => seq for reproductive sequence else runif for alea}
}
\value{
a sample generated according to a four parameter GGD
}
\description{
Random generation for the four-parameter Generalized Gamma Distribution
}
\seealso{
\code{\link{dgg4}},  \code{\link{pgg4}}, \code{\link{qgg4}}, \code{\link{hgg4}}, \code{\link{chgg4}}

#' @examples
## a = 2, l = 3, c = 4 and mu = 5
rr = rgg4(100,2,3,4,5,sequence = TRUE)
c(mean(rr),sd(rr))
## [1] 7.5497902 0.3718677
}
