% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.R
\name{skewness}
\alias{skewness}
\title{Skewness of a data population (measure of symmetry)}
\usage{
skewness(data, type = "sample")
}
\arguments{
\item{data}{a sample (set of observations)}

\item{type}{a character string coding -- "sample" is computed from mean and standard deviation
of sample. "fisher" corresponds to fisher'skewness formula.}
}
\value{
S the skewness from a sample
}
\description{
Skewness of a data population (measure of symmetry)
}
