% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc.dgg4.R
\name{truncdgg4}
\alias{truncdgg4}
\title{Truncated probability density function of
four-parameter Generalized Gamma Distribution}
\usage{
truncdgg4(q, a, l, c, mu = 0, trt, log.p = FALSE)
}
\arguments{
\item{q}{vector of quantile}

\item{a}{scale parameter}

\item{l}{shape parameter}

\item{c}{power shape parameter}

\item{mu}{location parameter}

\item{trt}{vector of quantile for truncated pdf}

\item{log.p}{Logical; if TRUE, probabilities p are given as log(p) only if a,l and c are strictly positives}
}
\value{
probability
}
\description{
Truncated probability density function of
four-parameter Generalized Gamma Distribution
}
