% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_theo.R
\name{var_theo}
\alias{var_theo}
\title{Variance estimated from the Generalized Gamma four-parameters. the function
allows us to verify the results regarding parameters estimated with mimic function}
\usage{
var_theo(theta)
}
\arguments{
\item{theta}{vector of 4-parameters.}
}
\value{
variance from parameters
}
\description{
Variance estimated from the Generalized Gamma four-parameters. the function
allows us to verify the results regarding parameters estimated with mimic function
}
